Practice, create three endpoints

+++ sb-ecom-practice/src/main/java/com/ecommerce/project/model/Category.java
@@ -4,7 +4,7 @@ public class Category {
     private long categoryId;
     private String categoryName;
 
-    public Category(long categoryId, String categoryName) {
+    public Category(int categoryId, String categoryName) {
         this.categoryId = categoryId;
         this.categoryName = categoryName;
     }
